Refactor and clean up UTXO tests and benchmarks
- Removed the `record_codec_equivalence.rs` test file as it is no longer needed.
- Simplified `snapshot_roundtrip.rs` by removing redundant tests and assertions related to snapshot encoding and decoding.
- Updated documentation in `index-read-path.md` and `utxo-memory.md` to reflect the retirement of historical tests and clarify current benchmarks.
- Introduced `resolver_fallback.rs` to add contract tests for the index resolver's legacy and invalid-position fallbacks, ensuring proper resolution against the canonical block.
