+++ CONCEPTS.md
@@ -342,7 +342,13 @@ The former cost shape of every lossy-prefix index resolver: read the block once 
 The invariant that lets index row values carry transaction byte positions without a block tag. The valid-state prerequisite is strict: the single writer atomically commits every row mutation with its exact full-hash capability watermark, rolls rows back before a replacement, and snapshot queries accept a result only when every required watermark equals the applied tip and both revision and tip remain stable. The reader validates the complete position list before I/O: it must be nonempty, strictly increasing, unique, nonoverlapping, within the block bound, and free of arithmetic overflow. It then reads each range from the current canonical `(height, full hash)` and exact-checks the decoded txid or scripthash. If any position fails, the reader discards every tentative result for that row and scans the full block; it never skips one position and keeps the rest. This also handles ordinary 8-byte prefix collisions. A stale row under an accepted watermark requires manual mutation, broken backend atomicity, or storage corruption and is outside the valid index-state contract; a full scan cannot make an arbitrarily corrupted row key authoritative. A per-row block tag was rejected because it measured 0.66x on top of the 1.67x position cost. See *Prefix-row rescan cost* for the range-read gain and *Identity-bearing key* for the write-side rollback guard.
 
 ### Paired-arm benchmark
-A Criterion group holding the before and after implementations of one change over one identical fixture, so the ratio comes from a single run. Adopted because a stored baseline cannot be trusted across a rebuild, and because the before implementation is wanted anyway as the equivalence oracle — the same function serves both roles. Its second use is diagnostic: while both arms still call the same code, their spread *is* the harness noise floor, measured rather than assumed. That reading is what disqualified the 64-height `subscribe` and `get_balance` groups, whose identical arms differ by 2.0-2.4x on a laptop and therefore cannot resolve a 1.05x gate, while `get_history` held within 1%. A group that cannot resolve the gate does not get to report a win.
+A historical Criterion pattern holding the before and after implementations of
+one change over one identical fixture, so the ratio comes from a single run. It
+was useful while a refactor was under review because the before implementation
+also served as an independent equivalence oracle and exposed the harness noise
+floor. The retired before/oracle implementations and their equivalence tests
+are not current runtime or test contracts; retained benchmarks now measure
+shipped production paths and product-shaped workloads.
 
 ### Resolution-time sampling
 Recording a statistic when its outcome is known rather than when the subject arrives. The fee estimator counted a transaction against every confirmation target the moment it entered, so a fresh arrival was already a failure at every target and a burst silenced the estimator before anything had missed a deadline. It also broke the decay: the denominator had been decaying since entry while a confirmation arrived undecayed, reporting 81 successes in 100 as roughly 85%. Sampling numerator and denominator together at the moment a target resolves fixes both, because they then decay from the same block. The counterpart rule is that a subject leaving for an unrelated reason is untracked without being sampled: an eviction says something about the mempool, not about whether the transaction would have confirmed.
