+++ bin/bitcoin-rs/Cargo.toml
@@ -20,12 +20,6 @@ mimalloc.workspace = true
 anyhow.workspace = true
 bitcoin-rs-node.workspace = true
 
-[dev-dependencies]
-parking_lot.workspace = true
-serde_json.workspace = true
-sha2.workspace = true
-tempfile = "3"
-
 [features]
 default = ["rocksdb", "fjall", "redb", "mdbx", "kernel", "zmq"]
 rocksdb = ["bitcoin-rs-node/rocksdb"]
@@ -35,42 +29,3 @@ mdbx = ["bitcoin-rs-node/mdbx"]
 kernel = ["bitcoin-rs-node/kernel"]
 zmq = ["bitcoin-rs-node/zmq"]
 
-[[test]]
-name = "g01_headers_only_sync"
-path = "tests/gates/g01_headers_only_sync.rs"
-
-[[test]]
-name = "g03_kernel_parity"
-path = "tests/gates/g03_kernel_parity.rs"
-
-[[test]]
-name = "g04_consensus_vectors"
-path = "tests/gates/g04_consensus_vectors.rs"
-
-[[test]]
-name = "g06_snapshot_roundtrip"
-path = "tests/gates/g06_snapshot_roundtrip.rs"
-
-[[test]]
-name = "g07_storage_equivalence"
-path = "tests/gates/g07_storage_equivalence.rs"
-
-[[test]]
-name = "g10_reorg_deep"
-path = "tests/gates/g10_reorg_deep.rs"
-
-[[test]]
-name = "g11_crash_recovery"
-path = "tests/gates/g11_crash_recovery.rs"
-
-[[test]]
-name = "g12_graceful_shutdown"
-path = "tests/gates/g12_graceful_shutdown.rs"
-
-[[test]]
-name = "g13_lints_clean"
-path = "tests/gates/g13_lints_clean.rs"
-
-[[test]]
-name = "g15_workspace_version_sync"
-path = "tests/gates/g15_workspace_version_sync.rs"
