--- bin/bitcoin-rs/tests/gates/g06_snapshot_roundtrip.rs
@@ -1,15 +1,0 @@
-//! G6 — Snapshot round-trip.
-//! **G6 — Snapshot round-trip.** `bitcoin-rs --snapshot-dump /tmp/utxo.snap && bitcoin-rs --snapshot-load /tmp/utxo.snap` reproduces an identical UTXO set and coinstats hash. Format is `bitcoin-rs`'s own LE format (gocoin wire-compat dropped per ultrareview).
-
-#![allow(clippy::let_unit_value)]
-
-/// Gate G6 manual run instructions: run
-/// `cargo test -p bitcoin-rs --test g06_snapshot_roundtrip -- --ignored --nocapture`
-/// with a populated UTXO set, dump and load a snapshot, then compare the
-/// resulting UTXO set and coinstats hash. The in-memory path is covered by
-/// `crates/utxo` unit tests.
-#[test]
-#[ignore = "requires populated UTXO set; covered by `crates/utxo` unit tests for the in-memory path"]
-fn snapshot_round_trip() {
-    // Dump and reload a populated UTXO snapshot, then compare UTXO and coinstats roots.
-}
