+++ crates/mempool/Cargo.toml
@@ -32,8 +32,6 @@ thiserror.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
-proptest.workspace = true
-serde_json.workspace = true
 criterion.workspace = true
 
 [[bench]]
