+++ crates/primitives/Cargo.toml
@@ -21,5 +21,3 @@ thiserror.workspace = true
 tinyvec.workspace = true
 zerocopy.workspace = true
 
-[dev-dependencies]
-proptest.workspace = true
