+++ crates/utxo/tests/commit_roundtrip.rs
@@ -1,108 +1,12 @@
-//! Commit/get round-trip coverage for a synthetic UTXO set.
-use std::sync::Arc;
+//! Public commit/get coverage for the UTXO set.
 
 use bitcoin::{Amount, ScriptBuf};
 use bitcoin_rs_primitives::{Hash256, OutPoint, TxOut, varint};
 use bitcoin_rs_utxo::{
-    BlockChanges, UtxoAdd, UtxoChangeEvents, UtxoChangeListener, UtxoCommittedEvent, UtxoError,
-    UtxoInserted, UtxoKey, UtxoRemoved, UtxoSet, hash_serialized_3,
+    BlockChanges, UtxoAdd, UtxoError, UtxoSet, hash_serialized_3,
     set::{BorrowedBlockChanges, BorrowedUtxoAdd},
 };
-use parking_lot::Mutex;
 use sha2::{Digest, Sha256};
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-enum ListenerEvent {
-    InsertBatch(Vec<u32>),
-    Remove(u32),
-}
-
-#[derive(Clone, Debug)]
-struct RecordingListener {
-    events: Arc<Mutex<Vec<ListenerEvent>>>,
-}
-
-#[derive(Clone, Debug)]
-struct BatchRecordingListener {
-    events: Arc<Mutex<Vec<ListenerEvent>>>,
-    batch_calls: Arc<Mutex<usize>>,
-    operation_counts: Arc<Mutex<Vec<usize>>>,
-}
-
-fn record_insertions(events: &Arc<Mutex<Vec<ListenerEvent>>>, insertions: &[UtxoInserted<'_>]) {
-    events.lock().push(ListenerEvent::InsertBatch(
-        insertions
-            .iter()
-            .map(|insertion| insertion.op.vout)
-            .collect(),
-    ));
-}
-
-fn record_removals(events: &Arc<Mutex<Vec<ListenerEvent>>>, removals: &[UtxoRemoved]) {
-    events.lock().extend(
-        removals
-            .iter()
-            .map(|removal| ListenerEvent::Remove(removal.op.vout)),
-    );
-}
-
-fn record_event_batches(events: &Arc<Mutex<Vec<ListenerEvent>>>, batches: &[UtxoChangeEvents<'_>]) {
-    let mut events = events.lock();
-    for batch in batches {
-        batch.for_each(|event| match event {
-            UtxoCommittedEvent::InsertBatch(insertions) => {
-                events.push(ListenerEvent::InsertBatch(
-                    insertions
-                        .iter()
-                        .map(|insertion| insertion.op.vout)
-                        .collect(),
-                ));
-            }
-            UtxoCommittedEvent::RemoveBatch(removals) => {
-                events.extend(
-                    removals
-                        .iter()
-                        .map(|removal| ListenerEvent::Remove(removal.op.vout)),
-                );
-            }
-        });
-    }
-}
-
-impl UtxoChangeListener for RecordingListener {
-    fn on_insert_coins(&self, insertions: &[UtxoInserted<'_>]) {
-        record_insertions(&self.events, insertions);
-    }
-
-    fn on_remove_coins(&self, removals: &[UtxoRemoved]) {
-        record_removals(&self.events, removals);
-    }
-
-    fn on_committed_event_batches(&self, batches: &[UtxoChangeEvents<'_>]) {
-        record_event_batches(&self.events, batches);
-    }
-}
-
-impl UtxoChangeListener for BatchRecordingListener {
-    fn on_insert_coins(&self, insertions: &[UtxoInserted<'_>]) {
-        record_insertions(&self.events, insertions);
-    }
-
-    fn on_remove_coins(&self, removals: &[UtxoRemoved]) {
-        record_removals(&self.events, removals);
-    }
-
-    fn on_committed_event_batches(&self, batches: &[UtxoChangeEvents<'_>]) {
-        let mut batch_calls = self.batch_calls.lock();
-        *batch_calls = batch_calls.saturating_add(1);
-        drop(batch_calls);
-
-        self.operation_counts
-            .lock()
-            .extend(batches.iter().map(UtxoChangeEvents::operation_count));
-        record_event_batches(&self.events, batches);
-    }
-}
 
 fn txid(seed: u64) -> Hash256 {
     let mut bytes = [0_u8; 32];
@@ -189,28 +93,6 @@ fn borrowed_changes<'a>(
         changes.add(BorrowedUtxoAdd::new(*outpoint, txout, *coinbase, *height));
     }
     changes
-}
-
-#[test]
-fn commit_roundtrips_ten_thousand_outputs() -> Result<(), Box<dyn std::error::Error>> {
-    let set = UtxoSet::new();
-    let mut changes = BlockChanges::default();
-    let mut expected = Vec::with_capacity(10_000);
-
-    for i in 0_u64..10_000 {
-        let outpoint = OutPoint::new(txid(i), u32::try_from(i % 4)?);
-        let txout = txout(i);
-        expected.push((outpoint, txout.clone()));
-        changes.add(UtxoAdd::new(outpoint, txout, false, 100));
-    }
-
-    set.commit_block(&changes, &txid(10_001))?;
-
-    for (outpoint, txout) in expected {
-        assert_eq!(set.get(&outpoint), Some(txout));
-    }
-
-    Ok(())
 }
 
 #[test]
@@ -331,350 +213,6 @@ fn has_live_outputs_for_txid_tracks_any_remaining_vout() -> Result<(), Box<dyn s
 }
 
 #[test]
-fn listener_batches_inserts_without_crossing_overwrite_boundary()
--> Result<(), Box<dyn std::error::Error>> {
-    let events = Arc::new(Mutex::new(Vec::new()));
-    let mut set = UtxoSet::new();
-    set.set_listener(Box::new(RecordingListener {
-        events: Arc::clone(&events),
-    }));
-    let live_txid = txid(600);
-
-    let mut initial = BlockChanges::default();
-    initial.add(UtxoAdd::new(
-        OutPoint::new(live_txid, 0),
-        txout(600),
-        false,
-        300,
-    ));
-    initial.add(UtxoAdd::new(
-        OutPoint::new(live_txid, 1),
-        txout(601),
-        true,
-        300,
-    ));
-    set.commit_block(&initial, &txid(602))?;
-    events.lock().clear();
-
-    let mut overwrite = BlockChanges::default();
-    overwrite.add(UtxoAdd::new(
-        OutPoint::new(live_txid, 2),
-        txout(603),
-        false,
-        301,
-    ));
-    overwrite.add(UtxoAdd::new(
-        OutPoint::new(live_txid, 1),
-        txout(604),
-        false,
-        301,
-    ));
-    overwrite.add(UtxoAdd::new(
-        OutPoint::new(live_txid, 3),
-        txout(605),
-        true,
-        301,
-    ));
-    set.commit_block(&overwrite, &txid(606))?;
-
-    assert_eq!(
-        events.lock().clone(),
-        vec![
-            ListenerEvent::InsertBatch(vec![2]),
-            ListenerEvent::Remove(1),
-            ListenerEvent::InsertBatch(vec![1, 3]),
-        ]
-    );
-    assert_eq!(set.get(&OutPoint::new(live_txid, 1)), Some(txout(604)));
-    assert_eq!(set.get(&OutPoint::new(live_txid, 2)), Some(txout(603)));
-    assert_eq!(set.get(&OutPoint::new(live_txid, 3)), Some(txout(605)));
-    Ok(())
-}
-
-#[test]
-fn listener_emits_explicit_removes_before_add_batches_in_single_shard_commit()
--> Result<(), Box<dyn std::error::Error>> {
-    let events = Arc::new(Mutex::new(Vec::new()));
-    let mut set = UtxoSet::new();
-    set.set_listener(Box::new(RecordingListener {
-        events: Arc::clone(&events),
-    }));
-    let live_txid = txid(620);
-    let first = OutPoint::new(live_txid, 0);
-    let second = OutPoint::new(live_txid, 1);
-    let third = OutPoint::new(live_txid, 2);
-
-    let mut initial = BlockChanges::default();
-    initial.add(UtxoAdd::new(first, txout(620), false, 302));
-    initial.add(UtxoAdd::new(second, txout(621), false, 302));
-    set.commit_block(&initial, &txid(622))?;
-    events.lock().clear();
-
-    let mut mixed = BlockChanges::default();
-    mixed.remove(first);
-    mixed.add(UtxoAdd::new(third, txout(623), true, 303));
-    set.commit_block(&mixed, &txid(624))?;
-
-    assert_eq!(
-        events.lock().clone(),
-        vec![
-            ListenerEvent::Remove(0),
-            ListenerEvent::InsertBatch(vec![2])
-        ]
-    );
-    assert_eq!(set.get(&first), None);
-    assert_eq!(set.get(&second), Some(txout(621)));
-    assert_eq!(set.get(&third), Some(txout(623)));
-    Ok(())
-}
-
-#[test]
-fn listener_full_record_spend_preserves_remove_input_order()
--> Result<(), Box<dyn std::error::Error>> {
-    let events = Arc::new(Mutex::new(Vec::new()));
-    let mut set = UtxoSet::new();
-    set.set_listener(Box::new(RecordingListener {
-        events: Arc::clone(&events),
-    }));
-    let live_txid = txid(625);
-
-    let mut initial = BlockChanges::default();
-    for vout in 0_u32..4 {
-        initial.add(UtxoAdd::new(
-            OutPoint::new(live_txid, vout),
-            txout(u64::from(vout) + 625),
-            false,
-            303,
-        ));
-    }
-    set.commit_block(&initial, &txid(626))?;
-    events.lock().clear();
-
-    let mut spend = BlockChanges::default();
-    for vout in [3_u32, 0, 2, 1] {
-        spend.remove(OutPoint::new(live_txid, vout));
-    }
-    set.commit_block(&spend, &txid(627))?;
-
-    assert_eq!(
-        events.lock().clone(),
-        vec![
-            ListenerEvent::Remove(3),
-            ListenerEvent::Remove(0),
-            ListenerEvent::Remove(2),
-            ListenerEvent::Remove(1),
-        ]
-    );
-    assert!(!set.has_live_outputs_for_txid(&live_txid));
-    Ok(())
-}
-
-#[test]
-fn listener_batches_parallel_multi_shard_event_multiset() -> Result<(), Box<dyn std::error::Error>>
-{
-    let events = Arc::new(Mutex::new(Vec::new()));
-    let mut set = UtxoSet::new();
-    set.set_listener(Box::new(RecordingListener {
-        events: Arc::clone(&events),
-    }));
-
-    let mut initial = BlockChanges::default();
-    let mut removes = Vec::new();
-    let mut inserts = Vec::new();
-    for shard in 0_u8..20 {
-        let txid = txid_in_shard(shard, u64::from(shard) + 630);
-        assert_eq!(UtxoKey::from_txid(&txid).shard(), shard);
-        let remove = OutPoint::new(txid, u32::from(shard));
-        let insert = OutPoint::new(txid, u32::from(shard) + 100);
-        removes.push(remove);
-        inserts.push(insert);
-        initial.add(UtxoAdd::new(
-            remove,
-            txout(u64::from(shard) + 630),
-            false,
-            304,
-        ));
-    }
-    set.commit_block(&initial, &txid(632))?;
-    events.lock().clear();
-
-    let mut mixed = BlockChanges::default();
-    for index in (0..removes.len()).rev() {
-        mixed.add(UtxoAdd::new(
-            inserts[index],
-            txout(u64::try_from(index)? + 700),
-            false,
-            305,
-        ));
-        mixed.remove(removes[index]);
-    }
-    set.commit_block(&mixed, &txid(635))?;
-
-    let mut recorded = events
-        .lock()
-        .iter()
-        .flat_map(|event| match event {
-            ListenerEvent::InsertBatch(vouts) => vouts
-                .iter()
-                .copied()
-                .map(|vout| (true, vout))
-                .collect::<Vec<_>>(),
-            ListenerEvent::Remove(vout) => vec![(false, *vout)],
-        })
-        .collect::<Vec<_>>();
-    recorded.sort_unstable();
-    let mut expected = (0_u32..20)
-        .flat_map(|vout| [(false, vout), (true, vout + 100)])
-        .collect::<Vec<_>>();
-    expected.sort_unstable();
-    assert_eq!(recorded, expected);
-    for (index, (remove, insert)) in removes.iter().zip(&inserts).enumerate() {
-        assert_eq!(set.get(remove), None);
-        assert_eq!(set.get(insert), Some(txout(u64::try_from(index)? + 700)));
-    }
-    Ok(())
-}
-
-#[test]
-fn listener_batches_small_multi_shard_commits() -> Result<(), Box<dyn std::error::Error>> {
-    let events = Arc::new(Mutex::new(Vec::new()));
-    let batch_calls = Arc::new(Mutex::new(0_usize));
-    let operation_counts = Arc::new(Mutex::new(Vec::new()));
-    let mut set = UtxoSet::new();
-
-    let mut initial = BlockChanges::default();
-    let first_remove = OutPoint::new(txid_in_shard(0, 650), 0);
-    let second_remove = OutPoint::new(txid_in_shard(1, 651), 1);
-    let first_insert = OutPoint::new(txid_in_shard(0, 652), 100);
-    let second_insert = OutPoint::new(txid_in_shard(1, 653), 101);
-    initial.add(UtxoAdd::new(first_remove, txout(650), false, 306));
-    initial.add(UtxoAdd::new(second_remove, txout(651), false, 306));
-    set.commit_block(&initial, &txid(654))?;
-
-    set.set_listener(Box::new(BatchRecordingListener {
-        events: Arc::clone(&events),
-        batch_calls: Arc::clone(&batch_calls),
-        operation_counts: Arc::clone(&operation_counts),
-    }));
-    let mut mixed = BlockChanges::default();
-    mixed.add(UtxoAdd::new(second_insert, txout(653), false, 307));
-    mixed.remove(second_remove);
-    mixed.add(UtxoAdd::new(first_insert, txout(652), false, 307));
-    mixed.remove(first_remove);
-    set.commit_block(&mixed, &txid(655))?;
-
-    assert_eq!(*batch_calls.lock(), 1);
-    assert_eq!(operation_counts.lock().clone(), vec![2, 2]);
-    assert_eq!(
-        events.lock().clone(),
-        vec![
-            ListenerEvent::Remove(0),
-            ListenerEvent::InsertBatch(vec![100]),
-            ListenerEvent::Remove(1),
-            ListenerEvent::InsertBatch(vec![101]),
-        ]
-    );
-    assert_eq!(set.get(&first_remove), None);
-    assert_eq!(set.get(&second_remove), None);
-    assert_eq!(set.get(&first_insert), Some(txout(652)));
-    assert_eq!(set.get(&second_insert), Some(txout(653)));
-    Ok(())
-}
-
-#[test]
-fn same_txid_churn_preserves_live_outputs_and_record_shape()
--> Result<(), Box<dyn std::error::Error>> {
-    let set = UtxoSet::new();
-    let live_txid = txid(700);
-    let first = OutPoint::new(live_txid, 0);
-    let second = OutPoint::new(live_txid, 1);
-    let third = OutPoint::new(live_txid, 2);
-    let fourth = OutPoint::new(live_txid, 3);
-    let first_txout = txout(701);
-    let second_txout = txout(702);
-    let third_txout = txout(703);
-    let fourth_txout = txout(704);
-
-    let mut initial = BlockChanges::default();
-    initial.add(UtxoAdd::new(first, first_txout, false, 400));
-    initial.add(UtxoAdd::new(second, second_txout.clone(), false, 400));
-    initial.add(UtxoAdd::new(third, third_txout, false, 400));
-    set.commit_block(&initial, &txid(705))?;
-
-    assert_eq!(set.record_count(), 1);
-    assert_eq!(set.len(), 3);
-
-    let mut churn = BlockChanges::default();
-    churn.remove(first);
-    churn.remove(third);
-    churn.add(UtxoAdd::new(fourth, fourth_txout.clone(), true, 401));
-    set.commit_block(&churn, &txid(706))?;
-
-    assert_eq!(set.get(&first), None);
-    assert_eq!(set.get(&second), Some(second_txout));
-    assert_eq!(set.get(&third), None);
-    assert_eq!(set.get(&fourth), Some(fourth_txout));
-    assert!(set.has_live_outputs_for_txid(&live_txid));
-    assert_eq!(set.record_count(), 1);
-    assert_eq!(set.len(), 2);
-    Ok(())
-}
-
-#[test]
-fn borrowed_commit_matches_owned_commit_for_same_txid_churn()
--> Result<(), Box<dyn std::error::Error>> {
-    let owned_set = UtxoSet::new();
-    let borrowed_set = UtxoSet::new();
-    let live_txid = txid(7_700);
-    let first = OutPoint::new(live_txid, 0);
-    let second = OutPoint::new(live_txid, 1);
-    let third = OutPoint::new(live_txid, 2);
-    let fourth = OutPoint::new(live_txid, 3);
-    let second_txout = txout(7_702);
-    let fourth_txout = txout(7_704);
-    let initial_adds = vec![
-        (first, txout(7_701), false, 400),
-        (second, second_txout.clone(), false, 400),
-        (third, txout(7_703), false, 400),
-    ];
-    let borrowed_initial = borrowed_changes(&initial_adds, &[]);
-    let mut owned_initial = BlockChanges::default();
-    for (outpoint, txout, coinbase, height) in &initial_adds {
-        owned_initial.add(UtxoAdd::new(*outpoint, txout.clone(), *coinbase, *height));
-    }
-    owned_set.commit_block(&owned_initial, &txid(7_705))?;
-    borrowed_set.commit_borrowed_block(&borrowed_initial, &txid(7_705))?;
-
-    let removes = vec![first, third];
-    let churn_adds = vec![(fourth, fourth_txout.clone(), true, 401)];
-    let borrowed_churn = borrowed_changes(&churn_adds, &removes);
-    let mut owned_churn = BlockChanges::default();
-    for remove in &removes {
-        owned_churn.remove(*remove);
-    }
-    for (outpoint, txout, coinbase, height) in &churn_adds {
-        owned_churn.add(UtxoAdd::new(*outpoint, txout.clone(), *coinbase, *height));
-    }
-    owned_set.commit_block(&owned_churn, &txid(7_706))?;
-    borrowed_set.commit_borrowed_block(&borrowed_churn, &txid(7_706))?;
-
-    for set in [&owned_set, &borrowed_set] {
-        assert_eq!(set.get(&first), None);
-        assert_eq!(set.get(&second), Some(second_txout.clone()));
-        assert_eq!(set.get(&third), None);
-        assert_eq!(set.get(&fourth), Some(fourth_txout.clone()));
-        assert!(set.has_live_outputs_for_txid(&live_txid));
-        assert_eq!(set.record_count(), 1);
-        assert_eq!(set.len(), 2);
-    }
-    assert_eq!(
-        hash_serialized_3(&borrowed_set)?,
-        hash_serialized_3(&owned_set)?
-    );
-    Ok(())
-}
-
-#[test]
 fn borrowed_commit_preserves_invalid_add_atomicity() -> Result<(), Box<dyn std::error::Error>> {
     let set = UtxoSet::new();
     let retained = OutPoint::new(txid(8_010), 0);
@@ -708,54 +246,6 @@ fn borrowed_commit_preserves_invalid_add_atomicity() -> Result<(), Box<dyn std::
     );
     assert_eq!(set.get(&retained), Some(retained_txout));
     assert_eq!(set.get(&invalid_outpoint), None);
-    Ok(())
-}
-
-#[test]
-fn borrowed_commit_duplicate_vout_overwrite_matches_owned_listener_events()
--> Result<(), Box<dyn std::error::Error>> {
-    let events = Arc::new(Mutex::new(Vec::new()));
-    let mut set = UtxoSet::new();
-    set.set_listener(Box::new(RecordingListener {
-        events: Arc::clone(&events),
-    }));
-    let live_txid = txid(8_600);
-
-    let mut initial = BlockChanges::default();
-    initial.add(UtxoAdd::new(
-        OutPoint::new(live_txid, 0),
-        txout(8_600),
-        false,
-        300,
-    ));
-    initial.add(UtxoAdd::new(
-        OutPoint::new(live_txid, 1),
-        txout(8_601),
-        true,
-        300,
-    ));
-    set.commit_block(&initial, &txid(8_602))?;
-    events.lock().clear();
-
-    let borrowed_adds = vec![
-        (OutPoint::new(live_txid, 2), txout(8_603), false, 301),
-        (OutPoint::new(live_txid, 1), txout(8_604), false, 301),
-        (OutPoint::new(live_txid, 3), txout(8_605), true, 301),
-    ];
-    let borrowed = borrowed_changes(&borrowed_adds, &[]);
-    set.commit_borrowed_block(&borrowed, &txid(8_606))?;
-
-    assert_eq!(
-        events.lock().clone(),
-        vec![
-            ListenerEvent::InsertBatch(vec![2]),
-            ListenerEvent::Remove(1),
-            ListenerEvent::InsertBatch(vec![1, 3]),
-        ]
-    );
-    assert_eq!(set.get(&OutPoint::new(live_txid, 1)), Some(txout(8_604)));
-    assert_eq!(set.get(&OutPoint::new(live_txid, 2)), Some(txout(8_603)));
-    assert_eq!(set.get(&OutPoint::new(live_txid, 3)), Some(txout(8_605)));
     Ok(())
 }
 
@@ -1068,38 +558,6 @@ fn zero_and_unequal_script_lengths_roundtrip_and_scan() -> Result<(), Box<dyn st
     ];
     let expected_hash = expected_hash_serialized_3(&entries)?;
     assert_eq!(hash_serialized_3(&set)?, expected_hash);
-    Ok(())
-}
-
-#[test]
-fn owned_commit_repeated_same_vout_overwrite_in_single_run()
--> Result<(), Box<dyn std::error::Error>> {
-    let events = Arc::new(Mutex::new(Vec::new()));
-    let mut set = UtxoSet::new();
-    set.set_listener(Box::new(RecordingListener {
-        events: Arc::clone(&events),
-    }));
-    let live_txid = txid(830);
-    let op = OutPoint::new(live_txid, 0);
-
-    let txout1 = txout(830);
-    let txout2 = txout(831);
-    let txout3 = txout(832);
-
-    let mut changes = BlockChanges::default();
-    changes.add(UtxoAdd::new(op, txout1, false, 100));
-    changes.add(UtxoAdd::new(op, txout2, true, 101));
-    changes.add(UtxoAdd::new(op, txout3.clone(), false, 102));
-    set.commit_block(&changes, &txid(831))?;
-
-    assert_eq!(set.get(&op), Some(txout3.clone()));
-    let entry = set.get_entry(&op).ok_or("expected live output")?;
-    assert_eq!(entry.txout, txout3);
-    assert!(!entry.coinbase);
-    assert_eq!(entry.height, 102);
-
-    assert_eq!(set.record_count(), 1);
-    assert_eq!(set.len(), 1);
     Ok(())
 }
